requirements.txt
flask
nmap
scapy
pyserial
openai
vosk
plotly
opencv-python
pyyaml
opencv-python
pyttsx3
pyserial
flask
